requirements/requirements.txt
appdirs==1.4.3
astroid==2.3.3
atomicwrites==1.3.0
attrs==19.3.0
bcrypt==3.1.7
cached-property==1.5.1
cachetools==3.1.1
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
Click==7.0
colorama==0.4.1
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.4
cryptography==2.8
Django==2.2.7
djangorestframework==3.10.3
docker==4.1.0
docker-compose==1.25.0
dockerpty==0.4.1
docopt==0.6.2
drf-yasg==1.17.0
entrypoints==0.3
google-auth==1.7.1
googleapis-common-protos==1.6.0
idna==2.8
importlib-metadata==0.23
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10.3
jsonschema==3.2.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==7.2.0
mypy-extensions==0.4.3
packaging==19.2
paramiko==2.6.0
pathspec==0.6.0
pluggy==0.13.0
protobuf==3.10.0
psycopg2-binary==2.8.4
py==1.8.0
pyasn1==0.4.7
pyasn1-modules==0.2.7
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
PyJWT==1.7.1
pylint==2.4.4
pylint-django==2.0.13
PyNaCl==1.3.0
pyparsing==2.4.5
pypiwin32==223
pyrsistent==0.15.5
pytest==5.3.0
pytest-asyncio==0.10.0
pytest-cov==2.8.1
pytest-cover==3.0.0
pytest-coverage==0.0
pytz==2019.3
pywin32==227
PyYAML==5.1.2
regex==2019.11.1
requests==2.22.0
rsa==4.0
ruamel.yaml==0.16.5
ruamel.yaml.clib==0.2.0
six==1.13.0
sqlparse==0.3.0
texttable==1.6.2
toml==0.10.0
typed-ast==1.4.0
typing-extensions==3.7.4.1
uritemplate==3.0.0
urllib3==1.25.7
wcwidth==0.1.7
websocket-client==0.56.0
wrapt==1.11.2
zipp==0.6.0
